requirements.txt
jsonschema==2.6.0
PyYAML==3.12
emoji==0.4.5
ansicolors==1.1.8
expression-parser==0.0.4
coverage==4.4.2
pytest==3.3.1
pytest-cov==2.5.1
